+++ packages/serverless-fc-trigger/src/oss.ts
@@ -4,10 +4,9 @@ import { FCBaseTrigger } from './base';
  * https://help.aliyun.com/document_detail/62922.html
  */
 export class OSSTrigger extends FCBaseTrigger {
-
   handler;
 
-  async toArgs(): Promise<any []> {
+  async toArgs(): Promise<any[]> {
     const event = {
       events: [
         {
@@ -20,34 +19,35 @@ export class OSSTrigger extends FCBaseTrigger {
               arn: 'acs:oss:cn-shanghai:123456789:bucketname',
               name: 'testbucket',
               ownerIdentity: '123456789',
-              virtualBucket: ''
+              virtualBucket: '',
             },
             object: {
               deltaSize: 122539,
               eTag: '688A7BF4F233DC9C88A80BF985AB7329',
               key: 'image/a.jpg',
-              size: 122539
+              size: 122539,
             },
             ossSchemaVersion: '1.0',
-            ruleId: '9adac8e253828f4f7c0466d941fa3db81161e853'
+            ruleId: '9adac8e253828f4f7c0466d941fa3db81161e853',
           },
           region: 'cn-shanghai',
           requestParameters: {
-            sourceIPAddress: '140.205.128.221'
+            sourceIPAddress: '140.205.128.221',
           },
           responseElements: {
-            requestId: '58F9FF2D3DF792092E12044C'
+            requestId: '58F9FF2D3DF792092E12044C',
           },
           userIdentity: {
-            principalId: '123456789'
-          }
-        }
-      ]
+            principalId: '123456789',
+          },
+        },
+      ],
     };
 
     return [event, this.createContext()];
   }
-
 }
 
 export const oss = OSSTrigger;
+export const os = OSSTrigger;
+export const cos = OSSTrigger;